work on udbzone and nsec3. fix format warnings.  svn:NO TEST
git-svn-id: file:///svn/nsd/trunk@3250 a26ef69c-88ff-0310-839f-98b793d9c207

+++ tpkg/cutest/cutest_udb.c
@@ -193,7 +193,8 @@ void check_rptrs_cb(void* base, udb_rel_ptr* p, void* arg)
 	/* the pointer must be on the pointerlist of that chunk */
 	if(!find_ptr_on_chunk(base, p->data, UDB_SYSTOREL(base, p), udb)) {
 		printf("cannot find ptr %llu on its chunks list at %llu\n",
-			UDB_SYSTOREL(base, p), p->data);
+			(unsigned long long)UDB_SYSTOREL(base, p),
+			(unsigned long long)p->data);
 		abort();
 	}
 	CuAssertTrue(tc, find_ptr_on_chunk(base, p->data,
